requirements.txt
aio-pika==9.4.3
aiofiles==23.2.1
aiogram==3.11.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiormq==6.8.0
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
APScheduler==3.10.4
asgiref==3.8.1
asyncmy==0.2.9
attrs==23.2.0
billiard==4.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crontab==1.0.1
cryptography==43.0.0
freezegun==1.5.1
frozenlist==1.4.1
greenlet==3.0.3
httpio==0.3.0
idna==3.7
kombu==5.4.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
pamqp==3.3.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
telebot==0.0.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
